requirements.txt
# Core dependencies for AI-Driven Literature Review System

# Configuration and data handling
pyyaml==6.0.1
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2

# LLM providers and interfaces
openai==1.6.1
anthropic==0.8.1
google-generativeai==0.3.2
langchain==0.1.0
langchain-anthropic==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6

# Academic APIs and data sources
requests==2.31.0
aiohttp==3.9.1
beautifulsoup4==4.12.3
semanticscholar==0.5.0

# PDF and document processing
pdfplumber==0.10.3
pymupdf==1.23.8
python-docx==1.1.0
bibtexparser==1.4.1

# Text processing and analysis
nltk==3.8.1
scikit-learn==1.3.2
thefuzz==0.20.0

# Vector storage and retrieval
chromadb==0.4.22
sentence-transformers==2.2.2

# Utilities and error handling
tenacity==8.2.3
tqdm==4.66.1
python-dateutil==2.8.2
unidecode==1.3.8

# Testing and development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0

# Monitoring and logging
rich==13.7.0
loguru==0.7.2

# Optional: Local LLM support (uncomment if using local models)
# ollama==0.1.7
# transformers==4.36.2
# torch==2.1.2
# accelerate==0.25.0